requirements.txt
fastapi[standard]==0.116.1
pymongo==4.13.2
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1